machine-learning/crab_detection/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
albumentations>=1.3.1
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
transformers>=4.30.0
timm>=0.9.0
Pillow>=9.5.0
rfdetr
supervision>=0.16.0
